Integration of nodemailer-sendgrid-transport
I also refactored some of it such that it contains only the necessary
files, as well as follow the structure that the other sub-packages
follow.

This commit should help with the PRs:
- [NST#67](https://github.com/sendgrid/nodemailer-sendgrid-transport/pull/67)
- [NST#64](https://github.com/sendgrid/nodemailer-sendgrid-transport/pull/64)

And the following issues:
- [NST#65](https://github.com/sendgrid/nodemailer-sendgrid-transport/issues/65)
- [NST#25](https://github.com/sendgrid/nodemailer-sendgrid-transport/issues/25)

The problem I face despit this commit being here is that some
sub-packages require the old `sendgrid` package which works differently
then this one.